Tableau.vertical	R Documentation

Tableau de plusieurs lignes

Description

Présentation de tables sous forme de tableau de plusieurs lignes de données.

Usage

Tableau.vertical(colnames, rownames, extra = "", ...)

Arguments

colnames
Vecteur de nom de colonnes pour le tableau

rownames
Vecteur de nom de lignes pour le tableau

extra
(= "") Si la valeur de ce paramètre est "variation", la variation relative est calculée entre le début et la fin de la période en lignes

...
paramètres fonctionnels uniquement. Si un paramètre n'est pas une fonction, le tableau est vide dans son ensemble

Value

Base de données data.table mise en forme de tableau par la fonction knitr::kable comportant l'ensemble des paramètres mis en colonnes centrées, avec x comme noms de lignes.

Examples

Tableau(c("a", "b", "c", "d"), 1, 2, 3, 4)

| a | b | c | d |
|:-:|:-:|:-:|:-:|
| 1 | 2 | 3 | 4 |