altair	R Documentation

Altaïr: un paquet pour analyser les bases de paye XML.

Description

Altaïr: un paquet pour analyser les bases de paye XML.

Démographie

effectifs : tableau des effectifs, ETP et EQTP

pyramide des âges : de 15 à 68 ans, verticale, hommes/femmes.