essayer	R Documentation

Essaye d'exécuter une portion de code et en cas d'erreur continue l'exécution du script en renvoyant un message d'erreur non bloquant.

Description

Essaye d'exécuter une portion de code et en cas d'erreur continue l'exécution du script en renvoyant un message d'erreur non bloquant.

Usage

essayer(X, Y)

Arguments

code
Portion de code

message
Message d'erreur

Value

Valeur retournée par X en cas de succès, sinon objet de classe "try-error" retourné par try(code)

Examples

if (exists("e")) rm(e)
essayer({ a <- 1/e}, "division par inconnu")  # affichage du message d'erreur