net.eqtp	R Documentation

Tableau des rémunérations nettes moyennes par grade.

Description

Elabore un tableau des rémunérations nettes moyennes du personnel par grade et par année (hors SFT).

Usage

net.eqtp(Base = Paie, grade = NULL, classe = NULL, service = NULL,
  libellés = NULL, agr = FALSE, période = NULL, variation = FALSE,
  statut = NULL, catégorie = NULL, exclure.codes = NULL,
  quotité.nulle = FALSE)

Arguments

Base
Base des bulletins de paye, comportant pour l'ensemble de la période

les variables charactère suivantes :

Année

Matricule

Statut

Grade

les variables numériques :

quotité

réel entre 0 et 1

.

grade
Grade particulier. Tous les grades en l'absence de spécification.

classe
Liste de vecteurs de grades, chaque vecteur représentant une classe aggrégée, ou bien vecteur de chaîne de caractères représentant des expressions rationnelles sur les grades. Tous les grades en l'absence de spécification. La casse est ignorée pour les expressions rationnelles.

service
Services. Vecteur de chaînes de caractères exactes. Tous les services en l'absence de spécification.

libellés
Vecteur de libellés des agrégations de grades par expression régulière. Doit avoir la même dimension que le vecteur de regexp.

agr
Booléen (défaut FALSE). Si TRUE, l'expression régulière précédente conduit à agréger les grades décrits par le vecteur d'expressions régulières précédent : une ligne par composante du vecteur.

période
Vecteur des années considérées.

variation
Booléen Insérer une colonne des variations (défaut FALSE).

statut
Restreindre le tableau au vecteur des statuts en paramètres. Expressions exactes. Tous statuts par défaut.

catégorie
Catégorie statutaire (vecteur de lettres parmi 'A', 'B', 'C'). Par défaut A, B, C ou indéterminée.

exclure.codes
Codes de paye à exclure pour le calcul du coût salarial (vecteur de chaînes de caractères).

Value

Un tableau des effectifs mis en forme avec les grades en ligne et autant de colonnes numériques que d'années de période, plus une colonne de libellés.

Examples

net.eqtp()